+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+Unreleased
+++++++++++
+* add PayuProvider.refund
+* update payment.captured_amount only when order is completed
+* subtract refunds from payment.captured_amount rather than from payment.total
+
 1.2.3 (2022-01-25)
 ++++++++++++++++++
 * better distinct PayU API errors